+++ sem1/zuev/4_prj_1var_cpp/formula_Bine.cpp
@@ -1,12 +1,12 @@
 #include <iostream>
 #include <cmath>
 
-int formulaBine(int n)
+double formulaBine(int n)
 {
-    int fn;
+    double fn;
     const double FI = (1 + sqrt(5)) / 2;
     const double PSI = (1 - sqrt(5)) / 2;
-    
+
     fn = (pow(FI, n) - pow(PSI, n)) / sqrt(5);
 
     return fn;
